Update 00_Group2_Ent_Struc.sql
Added code for creating Groups table and its constraints. Added "GO" command below the code for creating the Agents table.

Changes to the constraints that conflict with the Data Dictionary needed to be made, as the GroupsPage and GroupsEmail fields would not allow more than one row to contain a null value if they were set to Unique.